test: Start testing the Multi-precision integer implementation

+++ include/math/operations/Bitwise.hpp
@@ -3,5 +3,5 @@
 #include <cstdint>
 #include <vector>
 
-std::vector<uint32_t> bitwiseAnd(const std::vector<uint32_t>& lhs, const std::vector<uint32_t>& rhs);
-std::vector<uint32_t> bitwiseOr(const std::vector<uint32_t>& lhs, const std::vector<uint32_t>& rhs);
+std::vector<uint64_t> bitwiseAnd(const std::vector<uint64_t>& lhs, const std::vector<uint64_t>& rhs);
+std::vector<uint64_t> bitwiseOr(const std::vector<uint64_t>& lhs, const std::vector<uint64_t>& rhs);
