+++ include/math/operations/Shifting.hpp
@@ -3,7 +3,7 @@
 #include <cstdint>
 #include <vector>
 
-std::vector<uint32_t> bitShiftRight(const std::vector<uint32_t>& digits, size_t count);
-std::vector<uint32_t> bitShiftLeft(const std::vector<uint32_t>& digits, size_t count);
-std::vector<uint32_t> digitShiftRight(const std::vector<uint32_t>& digits, size_t count);
-std::vector<uint32_t> digitShiftLeft(const std::vector<uint32_t>& digits, size_t count);
+std::vector<uint64_t> bitShiftRight(const std::vector<uint64_t>& digits, size_t count);
+std::vector<uint64_t> bitShiftLeft(const std::vector<uint64_t>& digits, size_t count);
+std::vector<uint64_t> digitShiftRight(const std::vector<uint64_t>& digits, size_t count);
+std::vector<uint64_t> digitShiftLeft(const std::vector<uint64_t>& digits, size_t count);
