+++ include/math/operations/Subtraction.hpp
@@ -3,4 +3,4 @@
 #include <cstdint>
 #include <vector>
 
-std::vector<uint32_t> subtractDigits(const std::vector<uint32_t>& minuend, const std::vector<uint32_t>& subtrahend);
+std::vector<uint64_t> subtractDigits(const std::vector<uint64_t>& minuend, const std::vector<uint64_t>& subtrahend);
