+++ src/math/operations/Multiplication.cpp
@@ -2,11 +2,10 @@
 
 #include <algorithm>
 
-#include "logging/Logger.hpp"
 #include "math/operations/Addition.hpp"
 #include "math/operations/Subtraction.hpp"
 
-std::vector<uint32_t> multiplyDigits(const std::vector<uint32_t>& multiplier, const std::vector<uint32_t>& multiplicand)
+std::vector<uint64_t> multiplyDigits(const std::vector<uint64_t>& multiplier, const std::vector<uint64_t>& multiplicand)
 {
 	size_t bits = std::max(multiplier.size(), multiplicand.size()) << 5;
 
@@ -15,7 +14,7 @@ std::vector<uint32_t> multiplyDigits(const std::vector<uint32_t>& multiplier, co
 	return karatsubaMultiplication(multiplier, multiplicand);  // Once a better algorithm is implemented limit karatsuba to 2048 digits.
 }
 
-std::vector<uint32_t> hardwareMultiplication(const std::vector<uint32_t>& multiplier, const std::vector<uint32_t>& multiplicand)
+std::vector<uint64_t> hardwareMultiplication(const std::vector<uint64_t>& multiplier, const std::vector<uint64_t>& multiplicand)
 {
 	uint64_t product = 1;
 
@@ -23,17 +22,17 @@ std::vector<uint32_t> hardwareMultiplication(const std::vector<uint32_t>& multip
 	if (!multiplicand.empty()) product *= multiplicand[0];
 
 	return {
-		static_cast<uint32_t>(product),
-		static_cast<uint32_t>(product >> 32),
+		static_cast<uint64_t>(product),
+		static_cast<uint64_t>(product >> 32),
 	};
 }
 
-std::vector<uint32_t> longMultiplication(const std::vector<uint32_t>& multiplier, const std::vector<uint32_t>& multiplicand)
+std::vector<uint64_t> longMultiplication(const std::vector<uint64_t>& multiplier, const std::vector<uint64_t>& multiplicand)
 {
 	size_t maxDigits = multiplier.size() + multiplicand.size();
 
 	// Fill product with zeros, so that it can be index (unlike with reserve)
-	std::vector<uint32_t> product(maxDigits, 0);
+	std::vector<uint64_t> product(maxDigits, 0);
 
 	for (size_t i = 0; i < multiplier.size(); ++i)
 	{
@@ -41,11 +40,11 @@ std::vector<uint32_t> longMultiplication(const std::vector<uint32_t>& multiplier
 		for (size_t j = 0; j < multiplicand.size(); ++j)
 		{
 			uint64_t prod = static_cast<uint64_t>(multiplier[i]) * static_cast<uint64_t>(multiplicand[j]) + product[i + j] + carry;
-			product[i + j] = static_cast<uint32_t>(prod);
+			product[i + j] = static_cast<uint64_t>(prod);
 			carry = prod >> 32;
 		}
 
-		product[i + multiplicand.size()] += static_cast<uint32_t>(carry);
+		product[i + multiplicand.size()] += static_cast<uint64_t>(carry);
 	}
 
 	// Remove leading zeros
@@ -54,7 +53,7 @@ std::vector<uint32_t> longMultiplication(const std::vector<uint32_t>& multiplier
 	return product;
 }
 
-std::vector<uint32_t> karatsubaMultiplication(const std::vector<uint32_t>& multiplier, const std::vector<uint32_t>& multiplicand)
+std::vector<uint64_t> karatsubaMultiplication(const std::vector<uint64_t>& multiplier, const std::vector<uint64_t>& multiplicand)
 {
 	if (multiplier.size() <= 1 && multiplicand.size() <= 1) return hardwareMultiplication(multiplier, multiplicand);
 
@@ -63,24 +62,24 @@ std::vector<uint32_t> karatsubaMultiplication(const std::vector<uint32_t>& multi
 	auto [lowMultiplier, highMultiplier] = splitVector(multiplier, m);
 	auto [lowMultiplicand, highMultiplicand] = splitVector(multiplicand, m);
 
-	std::vector<uint32_t> z0 = karatsubaMultiplication(lowMultiplier, lowMultiplicand);
-	std::vector<uint32_t> z1 = karatsubaMultiplication(addDigits(lowMultiplier, highMultiplier), addDigits(lowMultiplicand, highMultiplicand));
-	std::vector<uint32_t> z2 = karatsubaMultiplication(highMultiplier, highMultiplicand);
+	std::vector<uint64_t> z0 = karatsubaMultiplication(lowMultiplier, lowMultiplicand);
+	std::vector<uint64_t> z1 = karatsubaMultiplication(addDigits(lowMultiplier, highMultiplier), addDigits(lowMultiplicand, highMultiplicand));
+	std::vector<uint64_t> z2 = karatsubaMultiplication(highMultiplier, highMultiplicand);
 
 	return addDigits(addDigits((leftShiftDigits(z2, (m << 1))), (leftShiftDigits((subtractDigits(subtractDigits(z1, z2), z0)), m))), z0);
 }
 
-std::pair<std::vector<uint32_t>, std::vector<uint32_t>> splitVector(const std::vector<uint32_t>& vec, size_t index)
+std::pair<std::vector<uint64_t>, std::vector<uint64_t>> splitVector(const std::vector<uint64_t>& vec, size_t index)
 {
 	index = std::min(index, vec.size());
 
-	std::vector<uint32_t> lower(vec.begin(), vec.begin() + index);
-	std::vector<uint32_t> upper(vec.begin() + index, vec.end());
+	std::vector<uint64_t> lower(vec.begin(), vec.begin() + index);
+	std::vector<uint64_t> upper(vec.begin() + index, vec.end());
 
 	return {lower, upper};
 }
 
-std::vector<uint32_t> leftShiftDigits(const std::vector<uint32_t>& input, size_t shift)
+std::vector<uint64_t> leftShiftDigits(const std::vector<uint64_t>& input, size_t shift)
 {
 	if (input.empty() || shift == 0)
 	{
@@ -88,7 +87,7 @@ std::vector<uint32_t> leftShiftDigits(const std::vector<uint32_t>& input, size_t
 	}
 
 	// Create a new vector with `shift` zeros appended at the end
-	std::vector<uint32_t> result(input.size() + shift, 0);
+	std::vector<uint64_t> result(input.size() + shift, 0);
 	std::copy(input.begin(), input.end(), result.begin() + shift);
 
 	return result;
